// backend/models/reporte.model.ts
export interface Reporte {
    IdReporte: number;
    NoNomina: number;
    MotivoReporte: string;
    PersonaReporto: string;
    FechaReporte: Date;
    SancionAplicada: string;
    NotasTalentoHumano: string;
    EstadoReporte: boolean;
    Nombre: string;
    Apellidos: string;
    NombreDepartamento:string;
    NombrePuesto: string;
  }